feat(web_service): introduced web-service api for e.g. edge-modules
