+++ Cargo.toml
@@ -11,6 +11,8 @@ repository = "git@github.com:omnect/omnect-device-service.git"
 version = "0.17.1"
 
 [dependencies]
+actix-server = "2.3"
+actix-web = "4.5"
 anyhow = "1.0"
 async-trait = "0.1"
 azure-iot-sdk = { git = "https://github.com/omnect/azure-iot-sdk.git", tag = "0.11.11", features = [
